<!---
Thanks for filing a pull request 😄 ! 

Pull requests are only accepted from Tamr employees. If there is a change you would like to see, please file a GitHub Issue or contact your Tamr Professional Services representative.

Below is a checklist of things to do before it can be merged.
-->

# ↪️ Pull Request

<!---
Provide a general summary of the pull request here
Please look for any issues that this PR resolves and tag them in the PR.
-->


## ✔️ PR Todo

- [x] Add documentation
- [x] Add examples
- [x] Ensure you are following the practices in the [contributor guide](https://docs.google.com/document/d/1c_hhpDMhO0zN793Z_HFOz42Q4g7l1nPRVQ5pXOUsMJU/edit) and [coding_standards](https://datatamr.atlassian.net/wiki/spaces/FEKB/blog/2020/04/14/1137606830/Customer+Success+Coding+Standards)
  * [Use consistent and descriptive name conventions](https://docs.google.com/document/d/1c_hhpDMhO0zN793Z_HFOz42Q4g7l1nPRVQ5pXOUsMJU/edit#heading=h.477s36w4rds9) 
  * [Uses google-style docstrings](https://datatamr.atlassian.net/wiki/spaces/FEKB/blog/2020/04/14/1137606830/Customer+Success+Coding+Standards#Google-style-docstrings)
  * [Uses versioned API/client methods only](https://docs.google.com/document/d/1c_hhpDMhO0zN793Z_HFOz42Q4g7l1nPRVQ5pXOUsMJU/edit#heading=h.yv6df99fyrq2)
  * [Avoid global variables](https://datatamr.atlassian.net/wiki/spaces/FEKB/blog/2020/04/14/1137606830/Customer+Success+Coding+Standards#Avoid-global-variables)
  * [Avoid giant try/catch](https://datatamr.atlassian.net/wiki/spaces/FEKB/blog/2020/04/14/1137606830/Customer+Success+Coding+Standards#Avoid-giant-try-/-except)
  * [Don't use mutable default variables](https://datatamr.atlassian.net/wiki/spaces/FEKB/blog/2020/04/14/1137606830/Customer+Success+Coding+Standards#Don%E2%80%99t-use-mutable-default-arguments)
  * [Use a `main()` function in your scripts](https://datatamr.atlassian.net/wiki/spaces/FEKB/blog/2020/04/14/1137606830/Customer+Success+Coding+Standards#Use-a-%E2%80%9Cmain()%E2%80%9D--function-in-your-scripts)
- [x] Added/updated testing for this change (ensure you are testing any changes/new code!)
- [x] Update title to link to TBOX issue (i.e. start pull request title with TBOX-xx) 
- [x] Approval from first code-reviewer
- [ ] Approval from second code-reviewer (only required if a second reviewer has left comments)
- [x] Pass all checks


This PR improves creation time of the Graph from Tamr projects by taking an API call out of the inner recursive loop and instead saving all projects in memory then passing that variable through. Here are the times of the tests (online tests) of the Graph component prior to this fix:

```
[success] 14.47% tests/workflow/concurrent/test_graph.py::test_from_project_list: 9.1969s
[success] 10.19% tests/workflow/concurrent/test_graph.py::test_get_edges: 6.4743s
[success] 9.83% tests/workflow/concurrent/test_graph.py::test_downstream_nodes: 6.2444s
[success] 9.81% tests/workflow/concurrent/test_graph.py::test_predecessors: 6.2357s
[success] 9.46% tests/workflow/concurrent/test_graph.py::test_add_edges: 6.0130s
[success] 9.36% tests/workflow/concurrent/test_graph.py::test_get_projects_by_tier: 5.9486s
[success] 9.33% tests/workflow/concurrent/test_graph.py::test_successors: 5.9321s
[success] 9.23% tests/workflow/concurrent/test_graph.py::test_get_end_nodes: 5.8643s
[success] 9.18% tests/workflow/concurrent/test_graph.py::test_get_source_nodes: 5.8323s
[success] 9.14% tests/workflow/concurrent/test_graph.py::test_get_upstream_projects: 5.8065s
[success] 0.00% tests/workflow/concurrent/test_graph.py::test_from_edges: 0.0002s
```

and here they are after:

```
[success] 14.94% tests/workflow/concurrent/test_graph.py::test_from_project_list: 7.3341s
[success] 10.79% tests/workflow/concurrent/test_graph.py::test_downstream_nodes: 5.2969s
[success] 10.70% tests/workflow/concurrent/test_graph.py::test_get_source_nodes: 5.2502s
[success] 10.57% tests/workflow/concurrent/test_graph.py::test_predecessors: 5.1860s
[success] 10.50% tests/workflow/concurrent/test_graph.py::test_successors: 5.1530s
[success] 10.44% tests/workflow/concurrent/test_graph.py::test_get_projects_by_tier: 5.1232s
[success] 9.28% tests/workflow/concurrent/test_graph.py::test_get_edges: 4.5544s
[success] 9.28% tests/workflow/concurrent/test_graph.py::test_add_edges: 4.5530s
[success] 8.95% tests/workflow/concurrent/test_graph.py::test_get_end_nodes: 4.3924s
[success] 4.55% tests/workflow/concurrent/test_graph.py::test_get_upstream_projects: 2.2330s
[success] 0.00% tests/workflow/concurrent/test_graph.py::test_from_edges: 0.0002s

```

Even for this simple graph performance is sped up by 10-50%. This improvement scales non-linearly with graph complexity and so is likely to be even more performant in the real world