@clint-richardson 2022-12-19
good catch @skalish - I've updated the git history by rebasing off of `develop`. I don't understand why the check is failing now - the error seems to be related to something in how we are setting up the environment:

```
/opt/hostedtoolcache/Python/3.8.15/x64/lib/python3.8/site-packages/numpy/__init__.py:284: in __getattr__
    raise AttributeError("module {!r} has no attribute "
E   AttributeError: module 'numpy' has no attribute 'bool'
```

@RJSanders do you have any ideas?