@RJSanders 2022-12-19
Definitely an issue with the environment setup, since the same failure occurs when testing the develop branch itself.

Looking online, I noticed that people who had similar errors were sometimes able to resolve them by reinstalling numpy using pip, which makes me think that there's a package version incompatibility somewhere.