@RJSanders 2022-12-19
Looks like @zbpvarun-tamr's upping of the Pandas requirement to 1.1.5 has done the trick.