@clint-richardson 2022-12-20
@RJSanders do you mind reviewing again? Since the code history has changed the previous review has gone stale